- [x] closes #12737, 
closes #13191
 - [ ] tests added / passed
 - [x] passes ``git diff upstream/master | flake8 --diff``
 - [x] whatsnew entry

```
    before     after       ratio
  [2de2884 ] [4b352d9 ]
-  164.84ms     1.73ms      0.01  groupby.groupby_transform_dataframe.time_groupby_transform_dataframe
     4.44ms     3.77ms      0.85  groupby.groupby_transform_series.time_groupby_transform_series
     6.40ms     4.76ms      0.74  groupby.groupby_transform_series2.time_groupby_transform_series2
```