@chris-b1 2016-05-16
This also fixes (unless it was deliberate?) the issue of `SeriesGroupBy.transform` losing the Series's name, xref #9862