@jreback 2016-05-18
thanks @chris-b1 very nice PR as usual!